Move MullvadFileProvider to common module
