+++ android/lib/ui/resource/src/main/res/xml/provider_paths.xml
